+++ packages/tsup-configs/src/index.ts
@@ -1,3 +1,3 @@
 export * from "./build.ts";
 export * from "./dev.ts";
-export { type DefineConfigOptions } from "./defineConfigBase.ts";
+export { type DefineConfigOptions } from "./mergeConfigs.ts";
