+++ packages/tsup-configs/src/mergeConfigs.ts
@@ -2,7 +2,7 @@ import { defineConfig, type Options } from "tsup";
 
 export type DefineConfigOptions = Parameters<typeof defineConfig>[0];
 
-export function defineConfigBase(config: DefineConfigOptions | undefined, baseConfig: Options): ReturnType<typeof defineConfig> {
+export function mergeConfigs(config: DefineConfigOptions | undefined, baseConfig: Options): ReturnType<typeof defineConfig> {
     if (config === undefined || (Array.isArray(config) && config.length === 0)) {
         return defineConfig(baseConfig);
     }
